moves tunable class file from application to kruize layer folder
