+++ src/main/java/com/autotune/analyzer/kruizeLayer/ValidateKruizeLayer.java
@@ -15,11 +15,9 @@
  *******************************************************************************/
 package com.autotune.analyzer.kruizeLayer;
 
-import com.autotune.analyzer.application.Tunable;
 import com.autotune.analyzer.kruizeLayer.presence.LayerPresenceQuery;
 import com.autotune.analyzer.utils.AnalyzerConstants;
 import com.autotune.analyzer.utils.AnalyzerErrorConstants;
-import com.autotune.utils.KruizeSupportedTypes;
 
 import java.util.ArrayList;
 import java.util.HashMap;
