+++ src/main/java/com/autotune/analyzer/kruizeLayer/layers/ContainerLayer.java
@@ -15,7 +15,7 @@
  *******************************************************************************/
 package com.autotune.analyzer.kruizeLayer.layers;
 
-import com.autotune.analyzer.application.Tunable;
+import com.autotune.analyzer.kruizeLayer.Tunable;
 import com.autotune.common.trials.ContainerConfigData;
 import com.autotune.utils.KruizeConstants;
 import io.fabric8.kubernetes.api.model.Quantity;
